fix: move error page to pages directory
